создает схемы для юзера, рефакторит код

+++ model/contacts.js
@@ -3,15 +3,21 @@ const Contact = require("./schemas/contactShema");
 // const db = require("./db");
 // const normalizeId = require("./normalize");
 
-const listContacts = async () => {
-  const results = await Contact.find({});
+const listContacts = async (userId) => {
+  const results = await Contact.find({ owner: userId }).populate({
+    path: "owner",
+    select: "name email",
+  });
   return results;
 };
 
-const getContactById = async (id) => {
-  const result = await Contact.findOne({ _id: id });
-  console.log(result.id);
-  console.log(result._id);
+const getContactById = async (id, userId) => {
+  const result = await Contact.findOne({ _id: id, owner: userId }).populate({
+    path: "owner",
+    select: "name email",
+  });
+  // console.log(result.id);
+  // console.log(result._id);
   return result;
 };
 
@@ -20,17 +26,17 @@ const addContact = async (body) => {
   return result;
 };
 
-const updateContact = async (id, body) => {
+const updateContact = async (id, body, userId) => {
   const result = await Contact.findByIdAndUpdate(
-    { _id: id },
+    { _id: id, owner: userId },
     { ...body },
     { new: true }
   );
   return result;
 };
 
-const removeContact = async (id) => {
-  const result = await Contact.findByIdAndRemove({ _id: id });
+const removeContact = async (id, userId) => {
+  const result = await Contact.findByIdAndRemove({ _id: id, owner: userId });
   return result;
 };
